requirements.txt
aio-pika==9.6.2
aiormq==6.9.4
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.2.25
click==8.3.2
dnspython==2.8.0
email-validator==2.3.0
fast-depends==3.0.8
fastapi==0.135.3
fastapi-cli==0.0.24
fastapi-cloud-cli==0.16.1
fastar==0.10.0
faststream==0.6.7
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jinja2==3.1.6
librt==0.9.0
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
mypy==1.20.0
mypy-extensions==1.1.0
pamqp==3.3.0
pathspec==1.0.4
propcache==0.4.1
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.2
pydantic-settings==2.13.1
pygments==2.20.0
python-dotenv==1.2.2
python-multipart==0.0.26
pyyaml==6.0.3
rich==14.3.4
rich-toolkit==0.19.7
rignore==0.7.6
ruff==0.15.10
sentry-sdk==2.57.0
shellingham==1.5.4
sqlalchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
